requirements.txt
absl-py==2.2.2
aiobotocore==2.21.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
ansicolors==1.1.8
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.1
asgiref==3.8.1
asteroid-filterbanks==0.4.0
asttokens==3.0.0
asyncssh==2.20.0
atpublic==5.1
attr==0.3.1
attrs==25.3.0
audioread==3.0.1
av==14.3.0
awscli==1.38.1
azure-core==1.32.0
azure-storage-blob==12.25.1
billiard==4.2.1
black==25.1.0
bleach==5.0.1
boto==2.49.0
boto3==1.37.1
botocore==1.37.1
braceexpand==0.1.7
cachetools==5.5.2
celery==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
configobj==5.0.9
contourpy==1.3.1
cryptography==44.0.2
ctranslate2==4.6.0
cycler==0.12.1
Cython==3.0.12
cytoolz==1.0.1
dataclasses-json==0.6.7
datamodel-code-generator==0.26.1
datasets==3.5.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
Django==5.1.8
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.10.3
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
docopt==0.6.2
docutils==0.16
dpath==2.2.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
dtw-python==1.5.3
dulwich==0.22.8
dvc==3.59.1
dvc-data==3.16.9
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-s3==3.2.0
dvc-studio-client==0.21.0
dvc-task==0.40.2
editdistance==0.8.1
einops==0.8.1
email_validator==2.2.0
entrypoints==0.4
executing==2.2.0
expiringdict==1.2.2
Faker==37.1.0
fastapi==0.115.12
fastapi-cli==0.0.7
faster-whisper==1.1.0
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.18.0
flatbuffers==25.2.10
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
funcy==2.0
genson==1.3.0
get-annotations==0.1.2
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.11.4
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gradio==5.25.1
gradio_client==1.8.0
grandalf==0.8
greenlet==3.1.1
groovy==0.1.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.71.0
gto==1.7.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
humanfriendly==10.0
humansignal-drf-yasg==1.21.10.post1
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
ijson==3.3.0
importlib_metadata==8.6.1
inflect==5.6.2
inflection==0.5.1
iniconfig==2.1.0
intervaltree==3.1.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
isort==5.13.2
iterative-telemetry==0.0.10
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jiwer==3.1.0
jmespath==1.0.1
joblib==1.4.2
jsf==0.11.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.7.2
kaldi-io==0.9.8
kiwisolver==1.4.8
kombu==5.5.2
label-studio==1.16.0
label-studio-sdk==1.0.10
langchain==0.3.24
langchain-community==0.3.23
langchain-core==0.3.57
langchain-openai==0.3.15
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langsmith==0.3.41
launchdarkly-server-sdk==8.2.1
lazy_loader==0.4
lhotse==1.31.0
librosa==0.11.0
lightning==2.5.1
lightning-utilities==0.14.3
lilcom==1.8.1
llvmlite==0.44.0
lockfile==0.12.2
lxml==5.3.1
lxml_html_clean==0.4.1
Mako==1.3.9
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.3.1
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.2
nbformat==5.10.4
nemo-toolkit==2.2.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn==8.2.0.51
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvidia-pyindex==1.0.9
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.21.0
openai==1.70.0
openai-whisper==20240930
opentelemetry-api==1.31.1
optuna==4.2.1
ordered-set==4.0.2
orjson==3.10.16
packaging==24.2
pandas==2.2.3
papermill==2.6.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
pooch==1.8.2
primePy==1.3
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyboxen==1.3.0
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.9.1
pydantic_core==2.33.0
pydot==3.0.4
pydub==0.25.1
pygit2==1.17.0
Pygments==2.19.1
pygtrie==2.5.0
PyJWT==2.10.1
pyparsing==3.2.3
pyRFC3339==2.0.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.4
python-multipart==0.0.20
pytorch-lightning==2.5.1
pytorch-metric-learning==2.8.1
pytz==2022.7.1
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-mock==1.12.1
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.1
rpds-py==0.24.0
rq==1.16.2
rsa==4.7.2
rstr==3.2.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.11.5
rules==3.4
s3fs==2025.3.2
s3transfer==0.11.3
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
scmrepo==3.3.10
semantic-version==2.10.0
semver==3.0.4
sentencepiece==0.2.0
sentry-sdk==2.25.1
setuptools==78.1.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.40
sqlparse==0.5.3
sqltrie==0.11.2
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tldextract==5.1.3
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
toolz==1.0.0
torch==2.6.0
torch-audiomentations==0.12.0
torch-stft==0.1.4
torch_pitch_shift==1.2.5
torchaudio==2.6.0
torchmetrics==1.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
triton==3.2.0
typer==0.15.2
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
Unidecode==1.4.0
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
voluptuous==0.15.2
watchfiles==1.0.5
wcwidth==0.2.13
webdataset==0.2.111
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
wheel==0.40.0
whisper-timestamped==1.15.8
whisperx==3.3.2
wrapt==1.17.2
xmljson==0.2.1
xxhash==3.5.0
yandex-cloud-ml-sdk==0.9.1
yandexcloud==0.341.0
yarl==1.18.3
zc.lockfile==3.0.post1
zipp==3.21.0
zstandard==0.23.0
